+++ assets/app/app/reset-password/token/template.hbs
@@ -1,2 +1,1 @@
 {{outlet}}
-
