+++ assets/app/reset-password/index/route.js
@@ -25,7 +25,7 @@
 import Ember from 'ember';
 
 export default Ember.Route.extend({
-	setupController(controller) {
-      controller.set('model', this.store.createRecord('reset-password-token', {}));
-    }
+  setupController(controller) {
+    controller.set('model', this.store.createRecord('reset-password', {}));
+  }
 });
