+++ assets/app/reset-password/model.js
@@ -26,6 +26,4 @@ import DS from 'ember-data';
 
 export default DS.Model.extend({
   email: DS.attr('string'),
-  password: DS.attr('string'),
-  token: DS.attr('string'),
 });
