+++ assets/app/router.js
@@ -49,7 +49,7 @@ Router.map(function() {
       this.route('machine', { path: '/:machine_id' });
     });
     this.route('apps', function() {
-      this.route('app', { path: '/:app_id' });
+      this.route('app', { path: '/:app_id' }, function() {});
     });
     this.route('apps');
     this.route('files', function() {
@@ -68,12 +68,12 @@ Router.map(function() {
   this.route('login');
   this.route('direct-link');
   this.route('sign-up');
-  this.route('reset-password', function() {
-      this.route('reset-password-tokens', { path: '/:reset-password-token_id' });
-  });
 
   this.route('activate', function() {
     this.route('activate',  { path: '/:activate_id' });
+  });
+  this.route('reset-password', function() {
+    this.route('token',  { path: '/:reset-password_id' });
   });
 });
